+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased](https://github.com/JuptiterOne/graph-jira/compare/v1.9.8...HEAD)
 
+### Removed
+
+- Removed default `jira_user.email` property of `donotemail@example.com`
+
 ## [2.1.2] - 2021-08-12
 
 ### Added